+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - PR #257 Mark Cython-exposed C++ functions that raise
 - PR #269 Doc sync behavior in `copy_ptr_to_host`
 - PR #278 Allocate a `bytes` object to fill up with RMM log data
+- PR #280 Drop allocation/deallocation of `offset`
 
 ## Bug Fixes